Text Classification/ECB Crawler/requirements.txt
beautifulsoup4
selenium
webdriver-manager
pandas
requests
PyYAML
tqdm